server/requirements.txt
aiosmtplib==3.0.2
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.3.0
blinker==1.9.0
boto3==1.37.13
botocore==1.37.13
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cloudinary==1.42.2
colorama==0.4.6
cryptography==44.0.2
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.11
fastapi-mail==1.4.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
loguru==0.7.3
Mako==1.3.9
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.29.0
win32_setctime==1.2.0
